refactor(agentd): consolidate MSB_* env var parsing into AgentdConfig
Read all MSB_* environment variables once at startup via
AgentdConfig::from_env() and pass the typed config by reference to init,
agent, and session. Eliminates repeated env var reads and parsing
downstream and fixes an env-var race that caused PTY test flakes (the
dead ENV_USER fallback in resolve_requested_user could race with tests
that set/unset MSB_USER).

BlockRootSpec keeps the kind-based enum dispatch introduced in #548:
DiskImage { device, fstype } and OciErofs { lower, upper, upper_fstype },
parsed from kind=disk-image,... and kind=oci-erofs,... forms. init.rs
matches on the enum and calls mount_disk_image / mount_oci_erofs +
pivot_to_newroot.
