+++ crates/agentd/lib/network.rs
@@ -3,91 +3,43 @@
 //! Configures the guest network interface using ioctls and netlink, following
 //! the parameters from host.
 
-use std::net::{Ipv4Addr, Ipv6Addr};
-
-use crate::error::{AgentdError, AgentdResult};
-
-//--------------------------------------------------------------------------------------------------
-// Types
-//--------------------------------------------------------------------------------------------------
-
-/// Parsed `MSB_NET` specification.
-#[derive(Debug)]
-struct NetSpec<'a> {
-    iface: &'a str,
-    mac: [u8; 6],
-    mtu: u16,
-}
-
-/// Parsed `MSB_NET_IPV4` specification.
-#[derive(Debug)]
-struct NetIpv4Spec {
-    address: Ipv4Addr,
-    prefix_len: u8,
-    gateway: Ipv4Addr,
-    dns: Option<Ipv4Addr>,
-}
-
-/// Parsed `MSB_NET_IPV6` specification.
-#[derive(Debug)]
-struct NetIpv6Spec {
-    address: Ipv6Addr,
-    prefix_len: u8,
-    gateway: Ipv6Addr,
-    dns: Option<Ipv6Addr>,
-}
+use crate::config::{NetIpv4Spec, NetIpv6Spec, NetSpec};
+use crate::error::AgentdResult;
 
 //--------------------------------------------------------------------------------------------------
 // Functions
 //--------------------------------------------------------------------------------------------------
 
-/// Sets the guest hostname from `MSB_HOSTNAME`.
+/// Sets the guest hostname.
 ///
 /// Calls `sethostname()`, writes `/etc/hostname`, and provisions
 /// `/etc/hosts` with localhost aliases and the hostname entry.
-pub fn apply_hostname() -> AgentdResult<()> {
-    let hostname = match std::env::var(microsandbox_protocol::ENV_HOSTNAME) {
-        Ok(v) if !v.is_empty() => Some(v),
-        _ => None,
-    };
+pub(crate) fn apply_hostname(hostname: Option<&str>) -> AgentdResult<()> {
+    linux::write_hosts_file(hostname)?;
 
-    linux::write_hosts_file(hostname.as_deref())?;
-
-    if let Some(ref name) = hostname {
+    if let Some(name) = hostname {
         linux::set_hostname(name)?;
     }
 
     Ok(())
 }
 
-/// Applies network configuration from `MSB_NET*` environment variables.
+/// Applies network configuration.
 ///
 /// Always provisions loopback, even when no external network interface is
-/// requested. Missing `MSB_NET` is not an error (no networking requested).
-/// Parse failures and configuration failures are hard errors.
-pub fn apply_network_config() -> AgentdResult<()> {
+/// requested. Missing `net` is not an error (no networking requested).
+pub(crate) fn apply_network_config(
+    net: Option<&NetSpec>,
+    net_ipv4: Option<&NetIpv4Spec>,
+    net_ipv6: Option<&NetIpv6Spec>,
+) -> AgentdResult<()> {
     linux::configure_loopback()?;
 
-    let val = match std::env::var(microsandbox_protocol::ENV_NET) {
-        Ok(v) if !v.is_empty() => v,
-        _ => return Ok(()),
+    let Some(net) = net else {
+        return Ok(());
     };
 
-    let net = parse_net(&val)?;
-
-    // Parse optional IPv4 config.
-    let ipv4 = match std::env::var(microsandbox_protocol::ENV_NET_IPV4) {
-        Ok(v) if !v.is_empty() => Some(parse_net_ipv4(&v)?),
-        _ => None,
-    };
-
-    // Parse optional IPv6 config.
-    let ipv6 = match std::env::var(microsandbox_protocol::ENV_NET_IPV6) {
-        Ok(v) if !v.is_empty() => Some(parse_net_ipv6(&v)?),
-        _ => None,
-    };
-
-    linux::configure_interface(&net, ipv4.as_ref(), ipv6.as_ref())
+    linux::configure_interface(net, net_ipv4, net_ipv6)
 }
 
 fn hosts_file_contents(hostname: Option<&str>) -> String {
@@ -112,172 +64,6 @@ fn hosts_file_contents(hostname: Option<&str>) -> String {
     s
 }
 
-/// Parses `MSB_NET` value: `iface=NAME,mac=AA:BB:CC:DD:EE:FF,mtu=N`
-fn parse_net(val: &str) -> AgentdResult<NetSpec<'_>> {
-    let mut iface = None;
-    let mut mac = None;
-    let mut mtu = 1500u16;
-
-    for part in val.split(',') {
-        if let Some(v) = part.strip_prefix("iface=") {
-            iface = Some(v);
-        } else if let Some(v) = part.strip_prefix("mac=") {
-            mac = Some(parse_mac(v)?);
-        } else if let Some(v) = part.strip_prefix("mtu=") {
-            mtu = v
-                .parse()
-                .map_err(|_| AgentdError::Init(format!("invalid MTU: {v}")))?;
-        } else {
-            return Err(AgentdError::Init(format!("unknown MSB_NET option: {part}")));
-        }
-    }
-
-    let iface = iface.ok_or_else(|| AgentdError::Init("MSB_NET missing iface=".into()))?;
-    let mac = mac.ok_or_else(|| AgentdError::Init("MSB_NET missing mac=".into()))?;
-
-    Ok(NetSpec { iface, mac, mtu })
-}
-
-/// Parses `MSB_NET_IPV4` value: `addr=A.B.C.D/N,gw=A.B.C.D[,dns=A.B.C.D]`
-fn parse_net_ipv4(val: &str) -> AgentdResult<NetIpv4Spec> {
-    let mut address = None;
-    let mut prefix_len = None;
-    let mut gateway = None;
-    let mut dns = None;
-
-    for part in val.split(',') {
-        if let Some(v) = part.strip_prefix("addr=") {
-            let (addr, prefix) = parse_cidr_v4(v)?;
-            address = Some(addr);
-            prefix_len = Some(prefix);
-        } else if let Some(v) = part.strip_prefix("gw=") {
-            gateway = Some(
-                v.parse::<Ipv4Addr>()
-                    .map_err(|_| AgentdError::Init(format!("invalid IPv4 gateway: {v}")))?,
-            );
-        } else if let Some(v) = part.strip_prefix("dns=") {
-            dns = Some(
-                v.parse::<Ipv4Addr>()
-                    .map_err(|_| AgentdError::Init(format!("invalid IPv4 DNS: {v}")))?,
-            );
-        } else {
-            return Err(AgentdError::Init(format!(
-                "unknown MSB_NET_IPV4 option: {part}"
-            )));
-        }
-    }
-
-    let address = address.ok_or_else(|| AgentdError::Init("MSB_NET_IPV4 missing addr=".into()))?;
-    let prefix_len =
-        prefix_len.ok_or_else(|| AgentdError::Init("MSB_NET_IPV4 missing addr=".into()))?;
-    let gateway = gateway.ok_or_else(|| AgentdError::Init("MSB_NET_IPV4 missing gw=".into()))?;
-
-    Ok(NetIpv4Spec {
-        address,
-        prefix_len,
-        gateway,
-        dns,
-    })
-}
-
-/// Parses `MSB_NET_IPV6` value: `addr=ADDR/N,gw=ADDR[,dns=ADDR]`
-fn parse_net_ipv6(val: &str) -> AgentdResult<NetIpv6Spec> {
-    let mut address = None;
-    let mut prefix_len = None;
-    let mut gateway = None;
-    let mut dns = None;
-
-    for part in val.split(',') {
-        if let Some(v) = part.strip_prefix("addr=") {
-            let (addr, prefix) = parse_cidr_v6(v)?;
-            address = Some(addr);
-            prefix_len = Some(prefix);
-        } else if let Some(v) = part.strip_prefix("gw=") {
-            gateway = Some(
-                v.parse::<Ipv6Addr>()
-                    .map_err(|_| AgentdError::Init(format!("invalid IPv6 gateway: {v}")))?,
-            );
-        } else if let Some(v) = part.strip_prefix("dns=") {
-            dns = Some(
-                v.parse::<Ipv6Addr>()
-                    .map_err(|_| AgentdError::Init(format!("invalid IPv6 DNS: {v}")))?,
-            );
-        } else {
-            return Err(AgentdError::Init(format!(
-                "unknown MSB_NET_IPV6 option: {part}"
-            )));
-        }
-    }
-
-    let address = address.ok_or_else(|| AgentdError::Init("MSB_NET_IPV6 missing addr=".into()))?;
-    let prefix_len =
-        prefix_len.ok_or_else(|| AgentdError::Init("MSB_NET_IPV6 missing addr=".into()))?;
-    let gateway = gateway.ok_or_else(|| AgentdError::Init("MSB_NET_IPV6 missing gw=".into()))?;
-
-    Ok(NetIpv6Spec {
-        address,
-        prefix_len,
-        gateway,
-        dns,
-    })
-}
-
-/// Parses a MAC address string like `02:5a:7b:13:01:02`.
-fn parse_mac(s: &str) -> AgentdResult<[u8; 6]> {
-    let mut mac = [0u8; 6];
-    let mut len = 0usize;
-    for (i, part) in s.split(':').enumerate() {
-        if i >= 6 {
-            return Err(AgentdError::Init(format!("invalid MAC address: {s}")));
-        }
-        mac[i] = u8::from_str_radix(part, 16)
-            .map_err(|_| AgentdError::Init(format!("invalid MAC octet: {part}")))?;
-        len = i + 1;
-    }
-    if len != 6 {
-        return Err(AgentdError::Init(format!("invalid MAC address: {s}")));
-    }
-    Ok(mac)
-}
-
-/// Parses an IPv4 CIDR like `100.96.1.2/30`.
-fn parse_cidr_v4(s: &str) -> AgentdResult<(Ipv4Addr, u8)> {
-    let (addr_str, prefix_str) = s
-        .split_once('/')
-        .ok_or_else(|| AgentdError::Init(format!("invalid IPv4 CIDR (missing /): {s}")))?;
-    let addr = addr_str
-        .parse::<Ipv4Addr>()
-        .map_err(|_| AgentdError::Init(format!("invalid IPv4 address: {addr_str}")))?;
-    let prefix = prefix_str
-        .parse::<u8>()
-        .map_err(|_| AgentdError::Init(format!("invalid IPv4 prefix length: {prefix_str}")))?;
-    if prefix > 32 {
-        return Err(AgentdError::Init(format!(
-            "IPv4 prefix length out of range (0-32): {prefix}"
-        )));
-    }
-    Ok((addr, prefix))
-}
-
-/// Parses an IPv6 CIDR like `fd42:6d73:62:2a::2/64`.
-fn parse_cidr_v6(s: &str) -> AgentdResult<(Ipv6Addr, u8)> {
-    let (addr_str, prefix_str) = s
-        .rsplit_once('/')
-        .ok_or_else(|| AgentdError::Init(format!("invalid IPv6 CIDR (missing /): {s}")))?;
-    let addr = addr_str
-        .parse::<Ipv6Addr>()
-        .map_err(|_| AgentdError::Init(format!("invalid IPv6 address: {addr_str}")))?;
-    let prefix = prefix_str
-        .parse::<u8>()
-        .map_err(|_| AgentdError::Init(format!("invalid IPv6 prefix length: {prefix_str}")))?;
-    if prefix > 128 {
-        return Err(AgentdError::Init(format!(
-            "IPv6 prefix length out of range (0-128): {prefix}"
-        )));
-    }
-    Ok((addr, prefix))
-}
-
 //--------------------------------------------------------------------------------------------------
 // Modules
 //--------------------------------------------------------------------------------------------------
@@ -285,9 +71,8 @@ fn parse_cidr_v6(s: &str) -> AgentdResult<(Ipv6Addr, u8)> {
 mod linux {
     use std::net::{Ipv4Addr, Ipv6Addr};
 
+    use crate::config::{NetIpv4Spec, NetIpv6Spec, NetSpec};
     use crate::error::{AgentdError, AgentdResult};
-
-    use super::{NetIpv4Spec, NetIpv6Spec, NetSpec};
 
     //----------------------------------------------------------------------------------------------
     // Types
@@ -330,14 +115,14 @@ mod linux {
     /// 7. Add IPv6 default route via netlink `RTM_NEWROUTE`
     /// 8. Write `/etc/resolv.conf`
     pub fn configure_interface(
-        net: &NetSpec<'_>,
+        net: &NetSpec,
         ipv4: Option<&NetIpv4Spec>,
         ipv6: Option<&NetIpv6Spec>,
     ) -> AgentdResult<()> {
-        let ifindex = get_ifindex(net.iface)?;
+        let ifindex = get_ifindex(&net.iface)?;
 
-        set_mac_address(net.iface, &net.mac)?;
-        set_mtu(net.iface, net.mtu)?;
+        set_mac_address(&net.iface, &net.mac)?;
+        set_mtu(&net.iface, net.mtu)?;
 
         if let Some(v4) = ipv4 {
             add_address_v4(ifindex, v4.address, v4.prefix_len)?;
@@ -346,7 +131,7 @@ mod linux {
             add_address_v6(ifindex, v6.address, v6.prefix_len)?;
         }
 
-        bring_interface_up(net.iface)?;
+        bring_interface_up(&net.iface)?;
 
         if let Some(v4) = ipv4 {
             add_default_route_v4(v4.gateway)?;
@@ -792,99 +577,6 @@ mod linux {
 #[cfg(test)]
 mod tests {
     use super::*;
-
-    #[test]
-    fn test_parse_net_full() {
-        let spec = parse_net("iface=eth0,mac=02:5a:7b:13:01:02,mtu=1500").unwrap();
-        assert_eq!(spec.iface, "eth0");
-        assert_eq!(spec.mac, [0x02, 0x5a, 0x7b, 0x13, 0x01, 0x02]);
-        assert_eq!(spec.mtu, 1500);
-    }
-
-    #[test]
-    fn test_parse_net_default_mtu() {
-        let spec = parse_net("iface=eth0,mac=02:00:00:00:00:01").unwrap();
-        assert_eq!(spec.mtu, 1500);
-    }
-
-    #[test]
-    fn test_parse_net_missing_iface() {
-        assert!(parse_net("mac=02:00:00:00:00:01").is_err());
-    }
-
-    #[test]
-    fn test_parse_net_missing_mac() {
-        assert!(parse_net("iface=eth0").is_err());
-    }
-
-    #[test]
-    fn test_parse_net_unknown_option() {
-        assert!(parse_net("iface=eth0,mac=02:00:00:00:00:01,bogus=42").is_err());
-    }
-
-    #[test]
-    fn test_parse_net_ipv4() {
-        let spec = parse_net_ipv4("addr=100.96.1.2/30,gw=100.96.1.1,dns=100.96.1.1").unwrap();
-        assert_eq!(spec.address, Ipv4Addr::new(100, 96, 1, 2));
-        assert_eq!(spec.prefix_len, 30);
-        assert_eq!(spec.gateway, Ipv4Addr::new(100, 96, 1, 1));
-        assert_eq!(spec.dns, Some(Ipv4Addr::new(100, 96, 1, 1)));
-    }
-
-    #[test]
-    fn test_parse_net_ipv4_no_dns() {
-        let spec = parse_net_ipv4("addr=10.0.0.2/24,gw=10.0.0.1").unwrap();
-        assert_eq!(spec.dns, None);
-    }
-
-    #[test]
-    fn test_parse_net_ipv4_missing_addr() {
-        assert!(parse_net_ipv4("gw=10.0.0.1").is_err());
-    }
-
-    #[test]
-    fn test_parse_net_ipv6() {
-        let spec = parse_net_ipv6(
-            "addr=fd42:6d73:62:2a::2/64,gw=fd42:6d73:62:2a::1,dns=fd42:6d73:62:2a::1",
-        )
-        .unwrap();
-        assert_eq!(
-            spec.address,
-            "fd42:6d73:62:2a::2".parse::<Ipv6Addr>().unwrap()
-        );
-        assert_eq!(spec.prefix_len, 64);
-        assert_eq!(
-            spec.gateway,
-            "fd42:6d73:62:2a::1".parse::<Ipv6Addr>().unwrap()
-        );
-        assert!(spec.dns.is_some());
-    }
-
-    #[test]
-    fn test_parse_mac_valid() {
-        let mac = parse_mac("02:5a:7b:13:01:02").unwrap();
-        assert_eq!(mac, [0x02, 0x5a, 0x7b, 0x13, 0x01, 0x02]);
-    }
-
-    #[test]
-    fn test_parse_mac_invalid() {
-        assert!(parse_mac("02:5a:7b").is_err());
-        assert!(parse_mac("zz:00:00:00:00:00").is_err());
-    }
-
-    #[test]
-    fn test_parse_cidr_v4() {
-        let (addr, prefix) = parse_cidr_v4("100.96.1.2/30").unwrap();
-        assert_eq!(addr, Ipv4Addr::new(100, 96, 1, 2));
-        assert_eq!(prefix, 30);
-    }
-
-    #[test]
-    fn test_parse_cidr_v6() {
-        let (addr, prefix) = parse_cidr_v6("fd42:6d73:62:2a::2/64").unwrap();
-        assert_eq!(addr, "fd42:6d73:62:2a::2".parse::<Ipv6Addr>().unwrap());
-        assert_eq!(prefix, 64);
-    }
 
     #[test]
     fn test_hosts_file_without_hostname() {
